+++ generate_programme.h
@@ -80,7 +80,7 @@ void verify_all_hand_made_cardinality_sequence_data (void);
 void verify_cardinality_sequence (int * sequence_data, int n);
 void display_digraph_node (aluminium_Christmas_tree * p, int n);
 void depth_first_search (aluminium_Christmas_tree * p, int n);
-long factorial (long n);
+long long factorial (long n);
 
 // These will only be needed until I get a proper generator written, but
 // they might be useful later as test cases for the generator. They have
